move instrumentation to common micronaut
