+++ dd-java-agent/instrumentation/micronaut/build.gradle
@@ -6,4 +6,5 @@ apply from: "$rootDir/gradle/java.gradle"
 
 dependencies {
   compileOnly group: 'io.micronaut', name: 'micronaut-http-server-netty', version: '2.0.0'
+  implementation project(':dd-java-agent:instrumentation:span-origin')
 }
